requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.4
sqlalchemy==2.0.32
databases==0.9.0
aiofiles==24.1.0
python-multipart==0.0.9
pydantic-settings==2.3.4
PyMuPDF==1.24.9
pdfplumber==0.11.4
ebooklib==0.18
docx2txt==0.8
beautifulsoup4==4.12.3
TTS==0.22.0
numpy==2.1.0
soundfile==0.12.1
